feat: add vite analyzer
