+++ packages/vite-service/package.json
@@ -34,6 +34,7 @@
     "vite": "^2.4.2"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.147"
+    "@types/lodash": "^4.14.147",
+    "rollup-plugin-visualizer": "^5.5.2"
   }
 }
